requirements.txt
django==3.0.6
psycopg2-binary
djangorestframework==3.11.0
beautifulsoup4
requests
